+++ src/main/java/pl/rscorporation/bookstoreapi/manager/BookService.java
@@ -3,15 +3,21 @@ package pl.rscorporation.bookstoreapi.manager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.context.event.EventListener;
+import org.springframework.data.domain.Pageable;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import pl.rscorporation.bookstoreapi.dao.AuthorRepository;
 import pl.rscorporation.bookstoreapi.dao.BookRepository;
+import pl.rscorporation.bookstoreapi.dao.dto.BookReadDTO;
+import pl.rscorporation.bookstoreapi.dao.dto.BookWriteDTO;
+import pl.rscorporation.bookstoreapi.dao.models.Author;
 import pl.rscorporation.bookstoreapi.dao.models.Book;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Optional;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class BookService {
@@ -30,19 +36,36 @@ public class BookService {
                 .orElseThrow(() -> new IllegalArgumentException("Author with given id not exists"));
     }
 
-    public Optional<Book> findById(Long id){
-        return bookRepository.findById(id);
+    public BookReadDTO findById(Long id){
+        Book toReturn = bookRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Book with this id not exists"));
+        return new BookReadDTO(toReturn);
     }
 
-    public List<Book> findAll(){
-        return bookRepository.findAll();
+    public List<BookReadDTO> findAll(){
+        return bookRepository.findAll().stream()
+                .map(book -> new BookReadDTO(book))
+                .collect(Collectors.toList());
     }
 
-    public Book save(Book book){
-        return bookRepository.save(book);
+    public List<BookReadDTO> findAll(Pageable page){
+        return bookRepository.findAll(page)
+                .stream()
+                .map(BookReadDTO::new)
+                .collect(Collectors.toList());
     }
 
-    public void deleteById(Long id){
+    public Book save(BookWriteDTO book){
+        Author bookAuthor = authorRepository.findById(book.getAuthorId())
+                .orElseThrow(() -> new IllegalArgumentException("Author with given id not exists"));
+
+        return bookRepository.save(book.toBook(bookAuthor));
+    }
+
+
+    public void deleteById(Long id)
+    {
+        if(!bookRepository.existsById(id))
+            throw new IllegalArgumentException("Book with given id not exists");
         bookRepository.deleteById(id);
     }
 }
